Push tests up in hierarchy so that we test both Ivy and Maven
